Replaced old viewModel with new viewModel
